+++ src/encoded/tests/test_annotated_filename.py
@@ -468,6 +468,8 @@ def test_get_donor_sex_and_age_parts(
         assert_filename_part_matches(result, expected, errors)
 
 
+SOME_FILE = {"data_category": ["Aligned Reads"]}
+REFERENCE_FILE = {"data_category": ["Reference Genome"]}
 SEQUENCER_CODE = "A"
 SOME_SEQUENCER = {"code": SEQUENCER_CODE}
 ANOTHER_SEQUENCER = {"code": "B"}
@@ -477,25 +479,27 @@ ANOTHER_ASSAY = {"code": "002"}
 
 
 @pytest.mark.parametrize(
-    "sequencers,assays,expected,errors",
+    "file,sequencers,assays,expected,errors",
     [
-        ([], [], "", True),
-        ([SOME_SEQUENCER], [], "", True),
-        ([], [SOME_ASSAY], "", True),
-        ([SOME_SEQUENCER], [SOME_ASSAY], f"{SEQUENCER_CODE}{ASSAY_CODE}", False),
-        ([SOME_SEQUENCER, ANOTHER_SEQUENCER], [SOME_ASSAY], "", True),
-        ([SOME_SEQUENCER], [SOME_ASSAY, ANOTHER_ASSAY], "", True),
-        ([SOME_SEQUENCER, SOME_ITEM], [SOME_ASSAY], "", True),
+        (SOME_FILE,[], [], "", True),
+        (SOME_FILE,[SOME_SEQUENCER], [], "", True),
+        (SOME_FILE,[], [SOME_ASSAY], "", True),
+        (SOME_FILE,[SOME_SEQUENCER], [SOME_ASSAY], f"{SEQUENCER_CODE}{ASSAY_CODE}", False),
+        (SOME_FILE,[SOME_SEQUENCER, ANOTHER_SEQUENCER], [SOME_ASSAY], "", True),
+        (REFERENCE_FILE,[SOME_SEQUENCER, ANOTHER_SEQUENCER], [SOME_ASSAY, ANOTHER_ASSAY], "XX", False),
+        (SOME_FILE,[SOME_SEQUENCER], [SOME_ASSAY, ANOTHER_ASSAY], "", True),
+        (SOME_FILE,[SOME_SEQUENCER, SOME_ITEM], [SOME_ASSAY], "", True),
     ],
 )
 def test_get_sequencing_and_assay_codes(
+    file: Dict[str, Any],
     sequencers: List[Dict[str, Any]],
     assays: List[Dict[str, Any]],
     expected: str,
     errors: bool,
 ) -> None:
     """Test sequencing and assay codes retrieval for annotated filenames."""
-    result = get_sequencing_and_assay_codes(sequencers, assays)
+    result = get_sequencing_and_assay_codes(file, sequencers, assays)
     assert_filename_part_matches(result, expected, errors)
 
 
@@ -630,14 +634,17 @@ def test_get_analysis(
     assert_filename_part_matches(result, expected, errors)
 
 
+SUPPLEMENTARY_FILE = {'@type': ["SupplementaryFile"]}
+OTHER_FILE = {'@type': ["VariantCals"]}
 @pytest.mark.parametrize(
-    "software,expected",
+    "file,software,expected",
     [
-        ([], ""),
-        ([{"version": "2.3.4"}], ""),
-        ([{"code": "foo", "version": "1.2.3"}], "foo_1.2.3"),
-        ([{"code": "foo", "version": "1.2.3"}, {"version": "2.3.4"}], "foo_1.2.3"),
+        (OTHER_FILE,[], ""),
+        (OTHER_FILE,[{"version": "2.3.4"}], ""),
+        (OTHER_FILE,[{"code": "foo", "version": "1.2.3"}], "foo_1.2.3"),
+        (OTHER_FILE,[{"code": "foo", "version": "1.2.3"}, {"version": "2.3.4"}], "foo_1.2.3"),
         (
+            OTHER_FILE,
             [
                 {"code": "foo", "version": "1.2.3"},
                 {"version": "2.3.4"},
@@ -645,13 +652,16 @@ def test_get_analysis(
             ],
             "bar_3.4.5_foo_1.2.3",
         ),
+        (SUPPLEMENTARY_FILE,[{"title": "Foo", "version": "1.2.3"}], "foo_1.2.3"),
     ],
 )
 def test_get_software_and_versions(
-    software: List[Dict[str, Any]], expected: str
+    file: Dict[str, Any],
+    software: List[Dict[str, Any]],
+    expected: str
 ) -> None:
     """Test software names and versions retrieval."""
-    result = get_software_and_versions(software)
+    result = get_software_and_versions(file ,software)
     assert result == expected