+++ CHANGELOG.rst
@@ -10,8 +10,9 @@ Change Log
 0.105.1
 =======
 `PR 277 DSA Release <https://github.com/smaht-dac/smaht-portal/pull/277>`_
-* In `commands/release-file.py` and `commands/create-annotated-filenames.py`, for DSA fasta files and chain files
-* Assay and sequencer codes value set to XX
+* In `commands/release-file.py` and `commands/create-annotated-filenames.py`:
+  * Assay and sequencer codes value set to XX for DSA fasta files and chain files
+  * For Supplementary Files, use `title` instead of `code` for software part of annotated filename
 
 0.105.0
 =======